utagms/requirements.txt
asgiref==3.7.2
backports.zoneinfo;python_version<"3.9"
Django==4.2.3
django-environ==0.10.0
djangorestframework==3.14.0
parameterized==0.9.0
psycopg2==2.9.6
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
django-cors-headers==4.2.0
uta-gms-engine==0.0.23